Restructuring the code
