+++ .github/workflows/auto-merge.yml
@@ -1,0 +1,39 @@
+name: Auto Merge PRs
+
+on:
+  pull_request:
+    types:
+      - labeled
+
+jobs:
+  auto-merge:
+    if: github.event.label.name == 'Merge'
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      pull-requests: write
+
+    steps:
+      - name: Check test status
+        uses: actions/github-script@v7
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          script: |
+            const pr = context.payload.pull_request;
+            const { data: checks } = await github.rest.checks.listForRef({
+              owner: context.repo.owner,
+              repo: context.repo.repo,
+              ref: pr.head.sha,
+            });
+
+            const failing = checks.check_runs.some(run => run.conclusion !== 'success');
+            if (failing) {
+              core.setFailed("RSpec checks have not passed. Auto-merge blocked.");
+            }
+
+      - name: Auto-merge PR
+        if: success() # only runs if the above step succeeds
+        uses: peter-evans/merge-pull-request@v3
+        with:
+          merge-method: squash
+          delete-branch: true
