+++ .github/workflows/rspec.yml
@@ -2,9 +2,9 @@ name: RSpec Tests
 
 on:
   push:
-    branches: [ main, develop ]
+    branches: [ main ]
   pull_request:
-    branches: [ main, develop ]
+    branches: [ main ]
 
 jobs:
   test:
@@ -17,7 +17,7 @@ jobs:
       - name: Set up Ruby
         uses: ruby/setup-ruby@v1
         with:
-          ruby-version: '3.2' # change this to match your project
+          ruby-version: '3.2'
           bundler-cache: true
 
       - name: Install dependencies
