+++ lib/string_calculator.rb
